+++ yorm/extended.py
@@ -4,7 +4,42 @@ import re
 
 import yaml
 
-from .standard import String
+from .standard import String, Integer, Float, Boolean
+from .container import Dictionary, List
+
+
+# standard types with None as a default #######################################
+
+
+class NoneString(String):
+
+    """Converter for the `str` type with `None` as default."""
+
+    DEFAULT = None
+
+
+class NoneInteger(Integer):
+
+    """Converter for the `int` type with `None` as default."""
+
+    DEFAULT = None
+
+
+class NoneFloat(Float):
+
+    """Converter for the `float` type with `None` as default."""
+
+    DEFAULT = None
+
+
+class NoneBoolean(Boolean):
+
+    """Converter for the `bool` type with `None` as default."""
+
+    DEFAULT = None
+
+
+# standard types with additional behavior #####################################
 
 
 class _Literal(str):
@@ -110,3 +145,54 @@ class Markdown(String):
             return Markdown.REGEX_SENTENCE_BOUNDARIES.sub('\n', stripped) + end
         else:
             return ''
+
+# container types with additional behavior ####################################
+
+
+class AttributeDictionary(Dictionary):
+
+    """Dictionary converter with keys available as attributes."""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.__dict__ = self
+
+    @classmethod
+    def default(cls):
+        """Create an uninitialized object with keys as attributes."""
+        if cls is AttributeDictionary:
+            msg = "AttributeDictionary class must be subclassed to use"
+            raise NotImplementedError(msg)
+
+        obj = cls.__new__(cls)
+        obj.__dict__ = obj
+        return obj
+
+
+class SortedList(List):
+
+    """List converter that is sorted on disk."""
+
+    @classmethod
+    def default(cls):
+        """Create an uninitialized object."""
+        if cls is SortedList:
+            msg = "SortedList class must be subclassed to use"
+            raise NotImplementedError(msg)
+        if not cls.item_type:
+            msg = "SortedList subclass must specify item type"
+            raise NotImplementedError(msg)
+
+        return cls.__new__(cls)
+
+    @classmethod
+    def to_data(cls, obj):
+        """Convert all attribute values for optimal dumping to YAML."""
+        value = cls.to_value(obj)
+
+        data = []
+
+        for item in sorted(value):
+            data.append(cls.item_type.to_data(item))
+
+        return data