+++ CHANGES.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+0.2.1 (2015-2-12)
+-----------------
+
+- Container types now extend their builtin type.
+- Added `None<Type>` extended types with `None` as a default.
+- Added `AttributeDictionary` with keys available as attributes.
+- Added `SortedList` that sorts when dumped.
+
 0.2 (2014-11-30)
 ----------------